requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.2
djangorestframework==3.14.0
dnspython==2.2.1
drf-yasg==1.21.4
et-xmlfile==1.1.0
eventlet==0.33.1
greenlet==2.0.0.post0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pip==22.3
prompt-toolkit==3.0.31
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
requests==2.28.1
ruamel.yaml.clib==0.2.7
ruamel.yaml==0.17.21
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.1
xlwt==1.3.0
